Image Widgets
Added the new standalone image widgets that don't share the same image. Each of them store a copy of its selected image. The picked image widget also stores a copy of the image chosen inside the app, so that it doesn't depend on the existence of the image inside the users gallery. Both types of widgets refresh when resized, in order to preserve the quality of the image.